requirements.txt
Flask
pandas
numpy
statsmodels
pmdarima
matplotlib
openai
python-dotenv